Fixing the meaning of "container_prefix"
Currently this API attribute means "exact name of a container" interface, but it was never the original intention.
It was just a temporary change when we have moved to the asynchronous provisioning method.
Even the name suggest this is just a prefix of the NICs :)
Functionally speaking this change was making it hard to attach multiple NICs to the same DanmNet... but nevermore!