Fixing iface deletion bug when netIds don't match
The project assumes that a DanmNet's NetworkID and ObjectMeta.Name parameters always match.
This is even outlined inthe schema descriptor.
However, recently we have introduced some changes which broke this assumption.
First, with the "default" feature, and then lately with the variable static delegate configuration.
The user might not even perceive it, but as a result of these changes interfaces weren't properly cleaned-up during CNI DEL invocations.
Instead of re-writing the whole codebase, we save the real name of the network as well into the interface's DanmEp, and use this parameter when we read the network descriptor during the deletion of the interface.
Meaning of NetworkID parameter left untouched in all the other places, but eventually, we need to revise when to sue which parameter.

Commit contains one more smal,, but important change: the first interface of a Pod will be always named "eth0" from now on.
This change will make it impossible for users to accidentally screw-up network naming, and thus kill their own Pod.