Adding the TenantNetwork specific code to the Webhook.
First, we need to create allocation bitarrays for every virtual interface when a TenantConfig is instantiated.
Then, when a TenantNetwork comes we need to:
- choose appropriate host_device, vlan, and vxlan IDs based on the configured TenantNetwork
- reserve the chosen VNI for the interface
- apply additional overwrites, such as NetworkType specific NetworkID

Existing validation rules were also enhanced.
We now allow the initial provisioning of host_device for TenantNetworks.
If user added it, we only match it against the Tenantconfig, and chose the configured VNIs, if, any.
We do the same for SR-IOV networks having DevicePools, instead of Devices.
Validation rules were also updated to make sure SR-IOV networks get special attention.