Added new event handler to Webhook.
This handler is responsible for handling DELETE events arriving for TenantNetworks.
During each such occurences we need to inspect if the TN being deleted had anyi VNI automatically allocated.
If yes, the VNI needs to be freed in the respective TenantConfig.

This hook can be used later on to add delete specific checks for other network APIs as well,
e.g. checking if Pods are still connected to a network being deleted etc.