Fixing bug of not properly cleaning up after ALL interfaces when SOME of them failed during create.
The root cause is that in some cases we have returned earlier during interface creation, than starting the asynch thread.
But if some other threads have already progressed to the point of forking, they were basically prematurely reaped by the early
termination of the whole process.
Nasty.

Resolution is getting rid of ALL early return cases. In case of an early failure we simply push the result in to the syncher object, even
if the error happened in the main thread.
Only Syncher can return from the main thread, no sooner and no later than all results are available: be them pushed from the main thread, or from a fork.