Netwatcher got new handlers for "Update" events
Whenever a network management API is updated in the API server, netwatcher will handle that event from now on.
It will take care of deleting the old, and creating the new interfaces on the host.

We only re-create the interfaces if something really changed: either VNI, or host_device.
We also stop the futile attempts of provisioning host interfaces for networks not needing them, such as SR-IOV networks.