Various corrections related to CNI delegation.
A bug is corrected both in cnidel, and metacni package related to Free-ing IPs when an error happened during the execution of the same asynch thread.
The correction is to immediately refresh the network manifest after allocation to avoid working on an out-of-date local copy.

CNI version support is now explicitly set to 0.3.1 to avoid any accidental mis-translation.

Started correctly filling IPAM config for static CNI deletes when DANM was reserving their IP during setup.