Adding IPv6 allocation pool to the API
This includes adapating the client code, and working it into the webhook validation rules.
Allocation bitmask for IPv6 becomes a mandatory attribute, just as for IPv4.
A network can only store 16M of addresses, total. Webhook makes sure this rule is never violated.
Real IPv6 allocation and freeing code is not yet changed by the PR.