Refactored IPv4 allocation function inside ipam package
The new function is generic enough to be re-used for IPv6 allocations as-is.
MAC address allocation removed from IPv6 allocations in preparations for the refactor.
It is however NOT removed from the APIs to maintain backward compatibility.
It will be simply deprecated instead.

Lastly change static IP allocation format to be a normal IP, i.e. without prefix.
Prefixed format continues to be accepted, again, for backward comp purposes.