Prefixing both V4 and V6 addresses with their network CIDR, rather than with their pool.
In case of V4 the two are the same of course.
In case of V6 prefixing with the pool CIDR could result in kernel throwing "no route to GW"
errors when we would like to add an otherwise valid IP route to a GW address which belongs to the network,
but falls outside the allocation pool.
Also the previous logic where we allowed static IPs to come from the network CIDR was accidentally changed.
This is now reverted, and made safe in case statc IPs fall outside of the pool.
If they are coming from the pool, we still reserve their place.
In a future enhancement we should read existing DanmEps to determine if the static IP was already assigned or not
when it is outside of the allocation pool.