Enhancing dummy interface capabilities with VLAN and PCI ID information.
When the VFIO bound VF is tagged with a VLAN, we create an extra VLAN IF on top of the dummy with the same tag.
The VF's PCI ID is added to the dummy (VLAN too) as an Alias.
This mapping should be enough for applications to map allocations to the proper device.