Adding obligatory cleanup to the beginning of every CNI ADD operation.
Before a CNI ADD is performed we unfortunetaly have to checkk if any DanmEps were already created for the exact same Pod.
This is because Kubelet can double invoke CNI ADDs without ever invoking CNI DEL in certain cases.
This results in unfortunate extra REST traffic, but is the only way to guarantee correct cluster state.