Refactoring post process handling to gracefully handle user space interfaces
The problem corrected in the first commit was just one symptom of ungraceful handling of user space interfaces.
The real, and generic correction is delivered by this commit.
Post-processing in general is now skipped for links not managed by the kernel.
A warning message is also written when this happens, so post-mortem debugging for kernel managed links is still possible.